Generalize code to be part of Node
